+++ app-client/src/__tests__/src/components/NavigationBar.test.tsx
@@ -9,7 +9,10 @@
 import * as React from 'react';
 import * as Adapter from 'enzyme-adapter-react-16';
 import {ShallowWrapper, shallow, configure, EnzymePropSelector} from 'enzyme';
-import {NavigationBar} from 'src/components/navigationBar/NavigationBar';
+import {RouteComponentProps} from 'react-router';
+import {NavigationBar, INavigationBarState} from 'src/components/navigationBar/NavigationBar';
+import {logOut, isLoggedIn} from 'src/utils/cognito';
+import {mockedHistory} from 'src/__tests__/testSetup.ts';
 import {
     Collapse,
     Navbar,
@@ -19,15 +22,15 @@ import {
     NavItem,
     NavLink,
 } from 'src/components/ReusableComponents';
-import {logOut, isLoggedIn} from 'src/utils/cognito';
-import {mockedHistory} from 'src/__tests__/testSetup.ts';
 
 const unroll = require('unroll');
 unroll.use(it);
 
 configure({adapter: new Adapter()});
 
-describe('NavigationBar tests.', (): void => {
+type ComponentTree = ShallowWrapper<RouteComponentProps<{email: string} | void>, INavigationBarState>;
+
+describe('NavigationBar tests.', () => {
 
     logOut = jest.fn<void>();
     isLoggedIn = jest.fn<boolean>()
@@ -38,24 +41,24 @@ describe('NavigationBar tests.', (): void => {
             return false;
         });
 
-    describe('When user is logged in', (): void => {
-        const componentTree: ShallowWrapper = shallow(
-            <NavigationBar history={mockedHistory}/>
-        );
+    const getComponentTree = (): ComponentTree => {
+        return shallow(<NavigationBar history={mockedHistory} />);
+    };
 
-        it('should log out when button is clicked', (): void => {
+    describe('When user is logged in', () => {
+        const componentTree: ComponentTree = getComponentTree();
+
+        it('should log out when button is clicked', () => {
             componentTree.find(NavLink).at(1).simulate('click');
             expect(logOut).toBeCalled();
             expect(mockedHistory.replace).toBeCalledWith('/');
         });
     });
 
-    describe('When user is logged out', (): void => {
-        const componentTree: ShallowWrapper = shallow(
-            <NavigationBar history={mockedHistory}/>
-        );
+    describe('When user is logged out', () => {
+        const componentTree: ComponentTree = getComponentTree();
 
-        it('should redirect to login page when button is clicked', (): void => {
+        it('should redirect to login page when button is clicked', () => {
             componentTree.find(NavLink).at(1).simulate('click');
             expect(mockedHistory.replace).toBeCalledWith('/');
         });
@@ -63,7 +66,7 @@ describe('NavigationBar tests.', (): void => {
         unroll('it should display #count #elementName elements', (
                 done: () => void,
                 args: {elementName: string, element: EnzymePropSelector, count: number}
-        ): void => {
+        ) => {
             expect(componentTree.find(args.element).length).toBe(args.count);
             done();
         }, [ // tslint:disable-next-line
@@ -77,9 +80,28 @@ describe('NavigationBar tests.', (): void => {
             ['NavLink', NavLink, 2],
         ]);
     
-        it('should set the state of isOpen to true', (): void => {
+        it('should set the state of isOpen to true', () => {
             componentTree.find(NavbarToggler).simulate('click');
             expect(componentTree.state().isOpen).toBeTruthy();
         });
+    });
+
+    describe('When the handleScroll is executed.', () => {
+
+        const componentTree: ComponentTree = getComponentTree();
+
+        unroll('it should set the background color to #color when the scrollY position is #scrollY', (
+                done: () => void,
+                args: {scrollY: number, color: string}
+        ) => {
+            Object.defineProperty(window, 'scrollY', {value: args.scrollY, writable: true});
+            componentTree.instance()[`handleScroll`]();
+            expect(componentTree.state().backgroundColor).toBe(args.color);
+            done();
+        }, [ // tslint:disable-next-line
+            ['scrollY', 'color'],
+            [100, 'white'],
+            [20, 'transparent'],
+        ]);
     });
 });