+++ app-client/src/__tests__/src/containers/homepage/AboutUs.test.tsx
@@ -8,23 +8,20 @@
 
 import * as React from 'react';
 import * as Adapter from 'enzyme-adapter-react-16';
+import {RouteComponentProps} from 'react-router';
 import {ShallowWrapper, shallow, configure, EnzymePropSelector} from 'enzyme';
-import {Homepage} from 'src/containers/homepage/Homepage';
+import {AboutUs} from 'src/containers/homepage/AboutUs';
 import {NavigationBar} from 'src/components/navigationBar/NavigationBar';
-import {Container} from 'src/components/ReusableComponents';
-
-const scrollParallax = require('react-scroll-parallax');
-const Parallax = scrollParallax.Parallax;
-const ParallaxProvider = scrollParallax.ParallaxProvider;
+import {PageContent} from 'src/components/PageContent';
 
 const unroll = require('unroll');
 unroll.use(it);
 
 configure({adapter: new Adapter()});
 
-describe('Homepage tests.', () => {
+describe('AboutUs tests.', () => {
 
-    const componentTree: ShallowWrapper = shallow(<Homepage />);
+    const componentTree: ShallowWrapper<RouteComponentProps<void>> = shallow(<AboutUs />);
 
     unroll('it should display #count #elementName elements', (
             done: () => void,
@@ -35,15 +32,9 @@ describe('Homepage tests.', () => {
     }, [ // tslint:disable-next-line
         ['elementName', 'element', 'count'],
         ['NavigationBar', NavigationBar, 1],
-        ['ParallaxProvider', ParallaxProvider, 1],
-        ['Container', Container, 2],
-        ['Parallax', Parallax, 1],
-        ['img', 'img', 5],
-        ['div', 'div', 9],
-        ['h1', 'h1', 1],
-        ['h3', 'h3', 3],
-        ['h4', 'h4', 1],
-        ['p', 'p', 6],
+        ['h2', 'h2', 1],
+        ['h5', 'h5', 1],
+        ['p', 'p', 5],
+        ['PageContent', PageContent, 1]
     ]);
-
 });